[CONFLICT] Adjust drag so it bounces
We bounce a little bit or invalidate if the user goes crazy on the bar.

It is not perfect, but should work enough for the moment.
You cannot drop on top of another task, you can drop close to a task and
adjust the start or end to match with a margin the previous task.

It should work enough for AMCELL

Signed-off-by: Sergio A. Martinez <smartinez@triditive.com>